fix(message): 修复获取未读消息总数接口返回值类型错误

+++ ctrl-blog-server/src/main/java/com/example/blog/modules/operation/service/MessageService.java
@@ -29,7 +29,7 @@ public interface MessageService extends IService<Message> {
      *
      * @return 未读消息数量
      */
-    Long getUnreadMessageCount();
+    Integer getUnreadMessageCount();
 
     /**
      * 标记单条消息为已读
